refactor: and memoize some chat components
