+++ src/pages/chat/components/conversation/ChatConversation.tsx
@@ -7,10 +7,9 @@ import {
 } from "@bosonprotocol/chat-sdk/dist/esm/util/v0.0.1/definitions";
 import { validateMessage } from "@bosonprotocol/chat-sdk/dist/esm/util/validators";
 import * as Sentry from "@sentry/browser";
-import dayjs from "dayjs";
 import { utils } from "ethers";
 import { useAccount } from "lib/utils/hooks/connection/connection";
-import { ArrowLeft, WarningCircle } from "phosphor-react";
+import { ArrowLeft } from "phosphor-react";
 import {
   ReactNode,
   useCallback,
@@ -19,31 +18,26 @@ import {
   useRef,
   useState
 } from "react";
-import InfiniteScroll from "react-infinite-scroll-component";
 import { useLocation } from "react-router-dom";
 import styled from "styled-components";
 
-import { Spinner } from "../../../../components/loading/Spinner";
 import InitializeChat from "../../../../components/modal/components/Chat/components/InitializeChat";
 import Grid from "../../../../components/ui/Grid";
-import SellerID from "../../../../components/ui/SellerID";
-import Typography from "../../../../components/ui/Typography";
 import { breakpoint } from "../../../../lib/styles/breakpoint";
 import { colors } from "../../../../lib/styles/colors";
 import { zIndex } from "../../../../lib/styles/zIndex";
 import { useInfiniteThread } from "../../../../lib/utils/hooks/chat/useInfiniteThread";
 import useCheckExchangePolicy from "../../../../lib/utils/hooks/offer/useCheckExchangePolicy";
 import { useBreakpoints } from "../../../../lib/utils/hooks/useBreakpoints";
-import { useDisputes } from "../../../../lib/utils/hooks/useDisputes";
 import { Exchange } from "../../../../lib/utils/hooks/useExchanges";
 import { useKeepQueryParamsNavigate } from "../../../../lib/utils/hooks/useKeepQueryParamsNavigate";
 import { useChatContext } from "../../ChatProvider/ChatContext";
 import { chatUrl } from "../../const";
 import { BuyerOrSeller, MessageDataWithInfo } from "../../types";
 import ExchangeSidePreview from "../ExchangeSidePreview";
-import Message from "../Message";
-import MessageSeparator from "../MessageSeparator";
 import { ChatConversationBottom } from "./ChatConversationBottom";
+import { Messages } from "./Messages";
+import { SellerComponent } from "./SellerComponent";
 
 const Container = styled.div`
   display: flex;
@@ -116,35 +110,6 @@ const Header = styled.div`
   }
 `;
 
-const Messages = styled.div<{ $overflow: string }>`
-  background-color: ${colors.lightGrey};
-  overflow: ${({ $overflow }) => $overflow};
-  display: flex;
-  flex-direction: column-reverse;
-  flex-grow: 1;
-`;
-
-const LoadingContainer = styled.div`
-  display: flex;
-  background-color: ${colors.lightGrey};
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  align-self: stretch;
-  gap: 0.25rem;
-`;
-
-const Conversation = styled.div<{ $alignStart: boolean }>`
-  display: flex;
-  flex-direction: column;
-  align-items: ${({ $alignStart }) =>
-    $alignStart ? "flex-start" : "flex-end"};
-  flex-grow: 1;
-  background-color: ${colors.lightGrey};
-  padding-bottom: 1.875rem;
-  position: relative;
-`;
-
 const SimpleMessage = styled.p`
   all: unset;
   display: block;
@@ -192,58 +157,6 @@ const NavigationMobile = styled.div`
   }
 `;
 
-const NoMoreMessages = styled.button.attrs({ type: "button" })`
-  background-color: ${colors.white};
-  color: ${colors.black};
-  border-radius: 1rem;
-  margin: 0.5rem;
-  padding: 0.2rem 0.7rem;
-  pointer-events: none;
-`;
-
-const LoadMoreMessages = styled.button.attrs({ type: "button" })`
-  background-color: ${colors.white};
-  color: ${colors.black};
-  border-radius: 1rem;
-  margin: 0.5rem;
-  padding: 0.2rem 0.7rem;
-  transition: all 300ms;
-
-  :hover {
-    background-color: ${colors.black};
-    color: ${colors.white};
-  }
-`;
-
-const SellerComponent = ({
-  size,
-  withProfileText,
-  exchange,
-  buyerOrSeller
-}: {
-  size: number;
-  withProfileText?: boolean;
-  exchange: Exchange | undefined;
-  buyerOrSeller: BuyerOrSeller;
-}) => {
-  if (!exchange) {
-    return null;
-  }
-  return (
-    <SellerID
-      offer={exchange?.offer}
-      buyerOrSeller={buyerOrSeller}
-      withProfileImage
-      accountImageSize={size}
-      withProfileText={withProfileText}
-    />
-  );
-};
-
-const getWasItSentByMe = (myAddress: string | undefined, sender: string) => {
-  return myAddress === sender;
-};
-
 type ChatConversationProps = {
   myBuyerId: string;
   mySellerId: string;
@@ -266,15 +179,6 @@ const ChatConversation = ({
   textAreaValue,
   refetchExchanges
 }: ChatConversationProps) => {
-  const { data: disputes } = useDisputes(
-    {
-      disputesFilter: {
-        exchange: exchange?.id
-      }
-    },
-    { enabled: !!exchange }
-  );
-  const dispute = disputes?.[0];
   const { account: address } = useAccount();
   const [hasError, setHasError] = useState<boolean>(false);
   const location = useLocation();
@@ -468,9 +372,6 @@ const ChatConversation = ({
     thread?.messages || []
   );
 
-  const hasMoreMessages = !isBeginningOfTimes;
-
-  const dataMessagesRef = useRef<HTMLDivElement>(null);
   const lastMessageRef = useRef<HTMLDivElement>(null);
   const scrollToBottom = useCallback(
     (scrollOptions: ScrollIntoViewOptions) =>
@@ -740,117 +641,18 @@ const ChatConversation = ({
       {canChat ? (
         <ContainerWithSellerHeader>
           <Messages
-            data-messages
-            ref={dataMessagesRef}
-            id="messages"
-            $overflow="auto"
-          >
-            <InfiniteScroll
-              inverse
-              next={loadMoreMessages}
-              hasMore={hasMoreMessages}
-              loader={<></>}
-              dataLength={thread?.messages.length || 0}
-              scrollableTarget="messages"
-              scrollThreshold="200px"
-              refreshFunction={loadMoreMessages}
-              pullDownToRefresh
-              pullDownToRefreshThreshold={50}
-              pullDownToRefreshContent={
-                <h3 style={{ textAlign: "center" }}>
-                  &#8595; Pull down to refresh
-                </h3>
-              }
-              releaseToRefreshContent={
-                <h3 style={{ textAlign: "center" }}>
-                  &#8593; Release to refresh
-                </h3>
-              }
-            >
-              <>
-                <LoadingContainer>
-                  {areThreadsLoading ? (
-                    <Spinner />
-                  ) : isBeginningOfTimes ? (
-                    <NoMoreMessages>No earlier messages</NoMoreMessages>
-                  ) : (
-                    <LoadMoreMessages onClick={() => loadMoreMessages()}>
-                      Load more messages
-                    </LoadMoreMessages>
-                  )}
-                </LoadingContainer>
-                {hasError && (
-                  <LoadingContainer>
-                    <WarningCircle color={colors.red} size={14} />
-                    <Typography color={colors.red}>
-                      There has been an error, please try again...
-                    </Typography>
-                  </LoadingContainer>
-                )}
-                {thread?.messages.map((message, index) => {
-                  const isFirstMessage = index === 0;
-                  const isPreviousMessageInADifferentDay = isFirstMessage
-                    ? false
-                    : dayjs(message.timestamp)
-                        .startOf("day")
-                        .diff(
-                          dayjs(thread.messages[index - 1].timestamp).startOf(
-                            "day"
-                          )
-                        ) > 0;
-                  const showMessageSeparator =
-                    isFirstMessage || isPreviousMessageInADifferentDay;
-                  const isLastMessage = index === thread.messages.length - 1;
-                  const wasItMe = getWasItSentByMe(address, message.sender);
-
-                  let buyerOrSeller: BuyerOrSeller;
-                  if (wasItMe) {
-                    if (iAmTheBuyer) {
-                      buyerOrSeller = exchange.buyer;
-                    } else {
-                      buyerOrSeller = exchange.seller;
-                    }
-                  } else {
-                    if (iAmTheBuyer) {
-                      buyerOrSeller = exchange.seller;
-                    } else {
-                      buyerOrSeller = exchange.buyer;
-                    }
-                  }
-                  const leftAligned = !wasItMe;
-                  const ref = isLastMessage ? lastMessageRef : null;
-                  return (
-                    <Conversation
-                      key={message.timestamp}
-                      $alignStart={leftAligned}
-                    >
-                      <>
-                        {showMessageSeparator && (
-                          <MessageSeparator message={message} />
-                        )}
-                        <Message
-                          exchange={exchange}
-                          dispute={dispute}
-                          message={message}
-                          isLeftAligned={leftAligned}
-                          lastReceivedProposal={lastReceivedProposal}
-                          lastSentProposal={lastSentProposal}
-                          ref={ref}
-                        >
-                          <SellerComponent
-                            size={32}
-                            withProfileText={false}
-                            exchange={exchange}
-                            buyerOrSeller={buyerOrSeller}
-                          />
-                        </Message>
-                      </>
-                    </Conversation>
-                  );
-                })}
-              </>
-            </InfiniteScroll>
-          </Messages>
+            hasError={hasError}
+            isBeginningOfTimes={isBeginningOfTimes}
+            exchange={exchange}
+            loadMoreMessages={loadMoreMessages}
+            thread={thread}
+            areThreadsLoading={areThreadsLoading}
+            address={address}
+            iAmTheBuyer={iAmTheBuyer}
+            lastMessageRef={lastMessageRef}
+            lastReceivedProposal={lastReceivedProposal}
+            lastSentProposal={lastSentProposal}
+          />
           <ChatConversationBottom
             proposal={lastReceivedProposal}
             isConversationBeingLoaded={isConversationBeingLoaded}
@@ -894,4 +696,5 @@ const ChatConversation = ({
     </Container>
   );
 };
+
 export default ChatConversation;
